Correcting Boolean default length handling
The default field length for boolean was not set, when rawAST descriptor
was generated for boolean types. The default length was set only for integer
and real at some places.

The fault was unnoticed because the RAW_encoder and decoder treated the 0 as
the default 1.
But when the raw offset/lenght was calculated by the compiler the boolean
fields was treated as zero length field, resulting a faulty offset for TAG fields